+++ CHANGELOG.md
@@ -9,16 +9,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fixed `--inline-threshold` deprecation warning in `embassy` examples.
 - Require safety documentation for unsafe blocks.
 - Fixed [FromRequest](https://docs.rs/picoserve/latest/picoserve/extract/trait.FromRequest.html) for `alloc::vec::Vec`, and by extension, `alloc::string::String`.
 
 ### Changes
 
-- `[picoserve::response::chunked::ChunkWriter::write_chunk](https://docs.rs/picoserve/latest/picoserve/response/chunked/struct.ChunkWriter.html)` no longer flushes the socket.
+- [`picoserve::response::chunked::ChunkWriter::write_chunk`](https://docs.rs/picoserve/latest/picoserve/response/chunked/struct.ChunkWriter.html) no longer flushes the socket.
 - Removed workaround for embassy-net TcpSocket::flush() never finishing due to upstream bugfix.
 - Tidied Content implementation of containers (e.g. Vec and String).
 - `impl<C: Content> IntoResponse for C`
+
+## [0.13.2] - 2024-12-10
+
+### Fixed
+
+- Fixed `--inline-threshold` deprecation warning in `embassy` examples.
 
 ### Added